Add aliasrc, shortcutrc, i3, nvim spelling and zsh
I added a bunch of new dotfiles to the mix. I now use zsh instead of
bash. In the process of switching to zsh, I have created aliasrc and
shortcutrc, two files that contain my shortcuts and aliases. I
enjoy this approach since it keeps aliases and shortcuts in sync for
zsh and bash.